develop compose singleton test
